consolidate getWyvernAsset, move bundle tests to own file
