+++ .gitignore
@@ -62,3 +62,5 @@ typings/
 
 # next.js build output
 .next
+
+.vscode
