+++ src/types.ts
@@ -191,7 +191,7 @@ export type WyvernAsset = WyvernNFTAsset | WyvernFTAsset
 
 // Abstractions over Wyvern assets for bundles
 export interface WyvernBundle {
-  assets: WyvernNFTAsset[]
+  assets: WyvernAsset[]
   name?: string
   description?: string
   external_link?: string
